Feature/send ping (#169)
* Implement sending ping frames.
* Check payload size of ping frames.

+++ src/sha1/sha1.c
@@ -233,10 +233,10 @@ void SHA1ProcessMessageBlock(SHA1Context *context)
      *  Initialize the first 16 words in the array W
      */
 	for (t = 0; t < 16; t++) {
-		W[t] = context->Message_Block[t * 4] << 24;
-		W[t] |= context->Message_Block[t * 4 + 1] << 16;
-		W[t] |= context->Message_Block[t * 4 + 2] << 8;
-		W[t] |= context->Message_Block[t * 4 + 3];
+		W[t] = (uint32_t)(context->Message_Block[t * 4]) << 24;
+		W[t] |= (uint32_t)(context->Message_Block[t * 4 + 1]) << 16;
+		W[t] |= (uint32_t)(context->Message_Block[t * 4 + 2]) << 8;
+		W[t] |= (uint32_t)(context->Message_Block[t * 4 + 3]);
 	}
 
 	for (t = 16; t < 80; t++) {
